Modules.PackageJson

Documentation

data Person

A person can be an author or a contributor.

Constructors

Person

Fields

personName :: String

personEmail :: String

personUrl :: String

data License

Use one of the common licenses or a custom one

Constructors

UNLICENSED

MIT

GPL

ISC

License Text

data Dependency

A JavaScript module the current one depends on. If it's installed from npm depUrl is Nothing.

Constructors

Dependency

Fields

depName :: String

depVersion :: String

depUrl :: Maybe String

getVersion :: Dependency -> String

dependencies :: [(String, String, Maybe String)] -> HashSet Dependency

data Script

Npm script which can be executed.

Constructors

Script

Fields

scriptName :: String

scriptBody :: String

scripts :: [(String, String)] -> HashSet Script

data Package

Constructors

Package

Fields

packName :: String

packVersion :: String

packDescription :: String

packMain :: String

packScripts :: HashSet Script

packKeywords :: [String]

packHomepage :: String

packLicense :: License

packPrivate :: Bool

packAuthor :: Person

packDependencies :: HashSet Dependency

packDevDependencies :: HashSet Dependency

getDepsMap :: (Package -> HashSet Dependency) -> Package -> HashMap String String

examplePerson :: Person

exampleDependency :: Dependency

exampleDevDependency :: Dependency

exampleScript :: Script

examplePackage :: Package

showPretty :: Package -> String

example :: String