Modules.Shakefile

Description

Documentation

data Command

Flag to change behavior of dockerBuildImage

Constructors

BuildOnly
Only build Docker images

BuildAndPush
Build and push to AWS

Utility Functions

clean :: Action ()

Delete all files from shake's cache

upgrade :: Action ()

Upgrade all Modules

yarnUpgrade :: FilePath -> Action ()

Upgrade yarn dependencies (changes `yarn.lock` file)

startsWithOneOf :: [String] -> String -> Bool

isModule :: String -> Bool

messageNoChange :: String -> [String] -> String

messageChange :: String -> [String] -> String

getDirsFromDiff :: String -> [String]

emptyToMaybe :: Foldable t => t a -> Maybe (t a)

Convert a possibly empty foldable to a maybe > emptyToMaybe "" == Nothing > emptyToMaybe "text" == Just "text"

ensureNonEmpty :: Foldable t => Maybe (t a) -> Maybe (t a)

Convert a possibly not existant and/or empty Foldable to a Maybe > ensureNonEmpty Nothing == Nothing > ensureNonEmpty (Just "") == Nothing > ensureNonEmpty (Just "text") == Just "text"

Build Targets

dockerBuildImage :: Command -> String -> String -> Action ()

Build of build & push a single docker image

dockerBuild :: Command -> Action ()

Build docker images of changed modules or all modules when other files were changed.

buildDocs :: Action ()

Generate documentation for Haskell code

customShakeOptions :: ShakeOptions

main :: IO ()

List all actions and call dockerBuild with BuildOnly per default